Moved methods and improved docs
