+++ tests/test_global_parallel_callback.py
@@ -1,10 +1,15 @@
+import doctest
+
 import joblib
 
 from tests._config_test import config, set_config
+from tpcp import parallel
 from tpcp.parallel import delayed, register_global_parallel_callback
 
 
 def test_simple_callback():
+    # We reset everything to avoid side effects
+    parallel._PARALLEL_CONTEXT_CALLBACKS = []
 
     set_config("set")
 
@@ -26,3 +31,19 @@ def test_simple_callback():
     register_global_parallel_callback(callback)
 
     assert joblib.Parallel(n_jobs=2)(delayed(func)() for _ in range(2)) == ["set", "set"]
+
+    # We reset everything to avoid side effects
+    parallel._PARALLEL_CONTEXT_CALLBACKS = []
+
+
+def test_doctest():
+    from tpcp import parallel
+
+    # We reset everything to avoid side effects
+    parallel._PARALLEL_CONTEXT_CALLBACKS = []
+
+    doctest_results = doctest.testmod(m=parallel)
+    assert doctest_results.failed == 0
+
+    # We reset everything to avoid side effects
+    parallel._PARALLEL_CONTEXT_CALLBACKS = []
