+++ tpcp/validate/_validate.py
@@ -9,8 +9,9 @@ from tqdm.auto import tqdm
 from tpcp import Dataset
 from tpcp._optimize import BaseOptimize
 from tpcp._utils._general import _aggregate_final_results, _normalize_score_results
-from tpcp._utils._multiprocess import TqdmParallel, delayed
+from tpcp._utils._multiprocess import TqdmParallel
 from tpcp._utils._score import _optimize_and_score
+from tpcp.parallel import delayed
 from tpcp.validate._scorer import _validate_scorer
